+++ app/src/main/java/co/id/roni/film_submission/favorite/movie/FavoriteMovieRepository.java
@@ -1,4 +1,4 @@
-package co.id.roni.film_submission.favorite;
+package co.id.roni.film_submission.favorite.movie;
 
 import android.app.Application;
 import android.os.AsyncTask;
