+++ app/src/main/java/co/id/roni/film_submission/favorite/movie/MovieFavoriteFragment.java
@@ -1,4 +1,4 @@
-package co.id.roni.film_submission.favorite;
+package co.id.roni.film_submission.favorite.movie;
 
 
 import android.os.Bundle;
