+++ app/src/main/res/layout/fragment_movie_favorite.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".favorite.MovieFavoriteFragment">
+    tools:context=".favorite.movie.MovieFavoriteFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <androidx.recyclerview.widget.RecyclerView
